Fixed up scanTag(). Tag scanned state was no longer needed. Switched to memcpy
The point of tag scanned state was to wait for user to remove the tag from the scanner. That isn't really necessary, so it was removed. tagScanTime also never got set, so that was fixed. Directly setting one array to another doesn't work, so memcpy is now being used instead. Tags are now only scanned while scanner is idle.